feat: Add created_by field to Sale model and update related files
The `created_by` field has been added to the `Sale` model to track the user who created the sale. This field is now included in the fillable array in the `Sale` model. The migration file `2024_05_13_144320_create_sales_table.php` has been updated to add the foreign key constraint for the `created_by` field. The `SaleManager` component has been modified to include the `created_by` field when creating a new sale. The `sales/index.blade.php` and `sales/show.blade.php` views have been updated to display the name of the user who created the sale.